Wstępny koniec backa przed testami

+++ Base of Taste/Base of Taste/Models/Skladnik.cs
@@ -16,6 +16,6 @@ namespace BazaSmakuAPI.Models
         public List<Alergen> Alergeny { get; set; }
         public List<ViewWarotscOdzywcza> WartoscOdzywcze { get; set; }
         public Jednostka Jednostka { get; set; }
-        public int Ilosc { get; set; }
+        public decimal Ilosc { get; set; }
     }
 }
